Add result view behaviour. Add travel confirm route. Fix bad redirect in authentication

+++ travels/models.py
@@ -10,7 +10,7 @@ DATETIME_FORMAT = '%d-%m-%Y %H:%M:%S'
 class Travel(models.Model):
     client=models.ForeignKey(Client, on_delete=models.CASCADE)
     driver=models.ForeignKey(Driver, on_delete=models.CASCADE)
-    fee=models.PositiveIntegerField(default=0)
+    fee=models.FloatField(default=0)
     start_date_time=models.DateTimeField(auto_now_add=True, blank=True)
     end_date_time=models.DateTimeField(auto_now_add=True, blank=True)
     start_pos=models.ForeignKey(PosLatLng, on_delete=models.CASCADE, related_name='startPos')
